Document how to copy PostgreSQL from one machine to another
I recently had to do this to copy the (correct) Mapit data from
staging and overwrite the equivalent (incorrect) data on
integration. It was fiddly and worth documenting in case we need
to do something similar again.
(We also don't have SCP / jumpbox proxy particularly well
documented at present; this guide should come in handy wherever
we need to upload or download something from an EC2.