Add Publishing API docs, App docs, DocumentTypes/Supertypes to search
- Moves this config into a new class 'ProxyPages' so that it can be tested
- Adds a 'title' frontmatter property so that it is indexed by middleman-search

This makes the search far more comprehensive and useful, although it does
slow the initial indexing down a bit, so the 'Loading search index'
message will stick around slightly longer. I think this is worth it for
the improved relevancy of search results, and if it becomes a problem we
can investigate a server-side search/index solution, or make better use
of caching. It should also only be a short term decrease in search speed;
when the latest version of middleman-search-gds is published, we'll
benefit by only indexing the titles of these external pages, not their
contents: https://github.com/alphagov/middleman-search/pull/5

Unfortunately, until that gem is released, I can't add the
govuk_schema_names pages as they break the indexing altogether.
It will be possible to add them in once the gem is updated.

Trello: https://trello.com/c/lVTGSjCA/142-imported-dev-docs-dont-appear-in-search