Use more inclusive terminology (public vs citizen)
"Citizen" is a very specific term and should only be used when we
specifically mean someone with a particular citizenship. Users of GOV.UK
services are generally not only British citizens but the broader public.
Using more general terms ("people", "public", "members of the public")
instead of "citizen" makes this clear and our docs more inclusive.