Add 'readme' method to GitHubRepoFetcher
We'll use this in subsequent commits to automatically pull in app
READMEs, which will be indexed in the dev docs search.
This should make it easier to surface important information.

Note that at this point, we're only retrieving the README as
markdown and still need to convert it to HTML. We'll tackle that
next.