Add 'parse' method to ExternalDoc
ExternalDoc is responsible for two things: fetching external
markdown, and parsing it as HTML according to our business
rules.

Prior to this commit, there was no way of simply passing
markdown to the ExternalDoc class. It makes sense to use
the same parsing functions whether the markdown is fetched
externally or passed inline.

This 'parse' method will be used when parsing the response
from the call to 'GitHubRepoFetcher.client.readme'.

Note that the 'AbsoluteLinkFilter' filter would not work
out of the box as it relies on a 'subpage_url' context.
This can be fixed up in future if desired.