Pull in application README to its corresponding app doc
This enriches each app doc page with its README content, which
is indexed by search and therefore makes our search more
powerful. It adds around 200KB to the search.json, which is
an acceptable amount for the functionality it provides.

I've added a warning that links might not work correctly,
as the content is pulled from an external repo and has
only been through a certain level of processing.