Restructure app doc pages
I wasn't keen on the mixture of markdown and HTML headings, so have
made that consistent (with the exception of the 'Links' heading,
which in its markdown form never gets converted from '## Links' to
the HTML equivalent. This is something that warrants further
investigation separately).

I've made the heading levels semantic, so that h1 is immediately
followed by h2 rather than h3. And I've consolidated some sections:

- Ownership and Dependencies now fall under one heading, and there
  is some basic logic to summarise it in a single sentence rather
  than take up too much screen real estate.
- SSH access is now a subsection of Hosting.
- 'Relevant manual pages' is now a subsection of Links.
- 'Example published pages' and 'Example rendered pages' now live
  more closely together.
- I've added the README at the end of the page