Import Email Alert API docs
This uses the same approach as the Publishing API. It's not
sustainable in the long term, but we plan on abstracting
the way we import docs in the future.

I've imported all the existing markdown files in the dir,
apart from those in the 'arch' subfolder, which don't seem
hugely important to import at this stage.