Update documentation for uploading HMRC files
Now explicitly uses gds-cli and govuk-connect for the SSH commands,
and documents some of the quirks around permissions (i.e. you need
to upload files to a home directory rather than directly to /tmp.
This may be because we have moved away from govukcli)

It also is more resilient in terms of how to connect to specific
machines. Where there are multiple machines of the same class,
the uploaded files won't have been uploaded to all of them, so
we should only be working with one (named) machine throughout.

It also makes it clearer that production _is_ your test
environment. Partners won't be able to test on staging or
integration, and there is a defined process for prefixing
with 'test-' on production.

Finally, it links to the content item publisher pages for the
pages that need republishing at the end, as they can be fiddly
to find, and notes that the version number in the page body
text should be updated to reflect the latest version
(this was missed on the Welsh version previously).