Update private repo instructions
We can use GitHub actions to do useful things like CI on short-lived private
repos. This probably means that enabling branch protection is worth it unless
it's for a very short-lived branch.

govuk-lint has been retired in favour of rubocop-govuk, so I've replaced that
instruction and put it first in the instructions because it's very quick compared
to the other steps.

I've added the assets:precompile step because we run this as part of CI currently
too.