Add troubleshooting tips for 'Smokey failed' alert
I considered putting this in its own alert document, but the
troubleshoot and remedy steps are almost identical to the
'Run high priority tests' alert - and I also struggled to
articulate the difference between the two errors - so have
kept them in the same document for now.
Hopefully someone else will come along and split them out
in the future if that would make the instructions easier
to follow.