Remove last_reviewed_on/review_in fields
Whilst this change happened on 29th October 2020, the commit date has been
backdated so that the 'last_updated' behaviour is unaffected. We don't want
to give a misleading impression that this document is more up to date than
it is because of the side-effect of removing some of its frontmatter.