Stop generating 'Icinga alerts' breadcrumb for non-existing anchor
There is no 'Icinga alerts' section on manual.html, as alerts are
their own special thing. We already have logic in the manual_layout
template to do something different with Icinga Alerts, so it doesn't
seem controversial to add another conditional.

I did consider removing all 'section: Icinga alerts' lines from the
docs under /alerts, as they are ALL the same and we can infer it
from the path. But that feels like a future optimisation. In the
meantime I just want to stop linking to an anchor point that
doesn't exist.