Add 'subsection' to Icinga Alerts documents
We currently have 87 different Icinga Alerts documents, and
currently no way of distinguishing between them. It's only
now I'm trying to audit our email-alert-related documentation
that I realise quite how many documents there are on the
subject and in an incident would be good to be able to find
quickly.

I see this as a first step in grouping related alerts
together and perhaps then exposing these on a top-level
'alerts.html' file in the same way as we do for 'manual.html'.
The 'subsection' concept may also be applicable to our general
documentation; for example:
> Databases
>> PostgreSQL
>> Backups
>> RDS
...etc.

But that is out of scope for this piece of work.