Move email troubleshooting tips into one place
This was already done in d000da0c62acf304e08a0d3c72144dc12f79740b, but
retrospectively in the wrong place. It seems odd that we had the
troubleshooting tips buried away inside documentation for a specific
alert, which all the other alerts linked to.

We now have a top-level 'Email troubleshooting' page, which is linked
from the manual.html, giving it more prominence and findability.
We can link to this page rather than to an anchor point of a semi
related page.