Explain how to work without GOV.UK Connect
Previously this manual was a mixture of using GOV.UK Connect for
SSH, and manual commands on the jumpbox. This rewrites the latter
part of the manual to instead focus on emulating GOV.UK Connect.
It should generally not be necessary to run manual commands on the
jumpboxes, since GOV.UK Connect provides wrapper for common ones.