Start indexing Schemas (but not their contents)
We should be able to search for a schema name and find the
associated schema in the dev docs. But like DocumentType
and Supertype, there's little value in indexing the contents
themselves, which would add an awful lot of bloat to the
search.json and slow down the speed of our search.

In fact, prior to middleman-search-gds v0.11.1, it was not
possible to include Schemas in the search as their contents
would be indexed by default and slow down the search so
much that it became unusable.