Rewrite errors manual to give the big picture
Previously I found this manual hard to read, because the policy at
the top only applied to two of the subsequent sections: most errors
are *not* reported to Sentry. Most of the linked URLs also did not
work for me, and seem largely unnecessary/verbose.

This rewrites the manual to surface the two approaches we have for
errors more concisely, while maintaining the original policy of the
associated RFC. Shortening this general manual about 'errors' will
also allow for a planned addition that describes what happens when
an application errors on GOV.UK (this will be in a future PR).