Fix rendering for content following 'RunRakeTask.links' invocation
Having the <ul><li>..</li></ul> on the same line interfered with the
markdown conversion of '## Heading' to '<h2>Heading</h2>'. We needed
the <ul> and </ul> to be on their own lines, and also to not have
any preceding white space (which is why I've used the '<<~' method
rather than '<<-').

Full conversation: https://github.com/alphagov/govuk-developer-docs/pull/2519\#discussion_r420917609
Co-authored-by: Ben Thorner <ben.thorner@digital.cabinet-office.gov.uk>