Documentation on Rails app conventions
This aims to be living documentation that captures the current patterns and
approaches GOV.UK uses in Rails applications. I'm expecting it'll only change
somewhat gradually, it feels like most of the things written in here could
have been catalogued a couple of years ago.

The motivation for putting this together was that it felt like some parts of
this were widely known among GOV.UK developers, others less so, but hardly any
parts of it are documented anywhere as a preference. This likely leads to
teams having to re-evaluate the same decisions' previous teams made. The
Publishing Workflow team went through a lot of this, which motivated me to
document some of those outcomes.

I've not been too specific on all the why's involved in these, mostly as this
would be quite verbose and often the decisions on the preferences are just
based on "this worked well for us, why do it differently?"