Improve ProxyPages spec tests
Rather than indexing the contents of application_template.html.md.erb,
which is more of a 'dashboard' and has little indexable value,
we pass an override of `content: ""` to only index by title,
which keeps the search light and fast.

I've now tweaked the tests to document this properly in code,
and also added a missing test for email_alert_api_docs.