Review 'error-reporting' manual
This refines the link to our SaaS config repo, to make it clearer
what we are referring to; the link is to a specific commit so that
a future reviewer can compare with any changes.

This also sets the review period to a year (we're unlikely to change
how we report errors), and adds a link to the manual from the newly
updated 'errors' manual.