Remove section about testing dashboards
We normally test all our changes, so we don't need to specify this.
Now that we no longer use development VMs, we should avoid suggesting
this method for testing, since people will be less likely to know
how to fix any issues they encounter with it. Deploying a branch to
Integration should be sufficient to test changes, which people are
generally familiar with from other apps.