Add documentation on how to add a Pingdom Check
This documentation is produced following a few meetings we've had about
the role Pingdom plays in our monitoring. I felt the most appropriate
place to capture this is in documentation for when you are adding a new
alert as that is the time when you need to consider these things.

Some follow-up work that pairs with this documentation is:

- We will be removing the checks in Puppet that checks are set, this is
  very out of date and there isn't an obvious way to ensure this stays
  up to date (ideally we'd instead automate populating Pingdom)
- We will remove a number of checks that seemed to not offer any
  additional external insights.