Remove 'apis/' folder
This will be re-added in the next commit via our new automated
config-based ingestion approach.

Anything two levels down in the docs/ folder will be lost, e.g.
alphagov/search-api's `docs/public-api/faceted-search.md` content.
For now I think it's fine to omit this level of granular detail;
we can look at doing a recursive import of docs/ in future if the
demand is there.

We also lose the fields reference for the search API which pulls
in the JSON from
https://github.com/alphagov/search-api/blob/master/config/schema/field_definitions.json

Again, this is a complexity I think we can live without for now;
we can link to that JSON file for the time being, and investigate
a JSON ingestion facility if needed in future.