Consolidate and rewrite deployment manuals
Previously we had several manuals about developing and deploying a
change to an app, with the 'development-pipeline' manual acting as
a kind of starting point. This consolidates all the manuals into a
single guide for how to raise and deploy a change, so that we have
less to review. Putting all the knowledge in one place will also
make it easier to add caveats for Continuous Deployment.