Review and restructure 'Reboot a machine'
I've reorganised this so that there is a section for 'Automatic rebooting'
which comes first, followed by a section dedicated to 'Manual rebooting'.
I've removed some of the duplication in the 'Rules of manual rebooting'
section, and made it a stronger warning that there are 'Special cases'
(a new heading for an existing section) that need to be considered for
certain machine classes.
I've moved the fabric script setup script to the top, and removed a broken
bit of markdown to do with the fabric script installation link.
And otherwise done a little bit of format fixing.