Set source_url for external docs
When we click 'View source', we expect it to take us to the main
document representing the page, so that we can easily edit the
information.
Until now, external docs' 'View source' link simply took you to
external_doc_template.html.erb, which is not a very useful
workflow.
We now bubble up the original source URL from the GitHub API
response so that we link to the right place.