Fix broken link due to markdown parse error
`[AWS S3 bucket] (read ["Assets: how they work"])`

Was being converted to:

`<a href="read [\"Assets: how they work\"]>AWS S3 bucket</a>`.

I.e. it was mistakingly treating the text following the first `]` as the URL.

We should look at why this parsing issue happens - whether it's down to
a bug in CommonMarker, or our markdown, or the dev docs.