Correct 'production_hosted_on' value for ephemeral apps
'govuk-related-links-recommender' and 'search-analytics' repos are
run at various intervals in Jenkins - they're not actually 'hosted'
anywhere. So the correct value for these is 'none'.

This makes the 'check_puppet_names' rake task happier, which was
rightly complaining that there is no corresponding puppet manifest
for these apps. (If they were hosted, we'd expect them to be
managed with puppet).