Fix failing 'check_puppet_names' rake task
The last app that this task was complaining about was 'licensify', as
it couldn't find licensify.pp in the default location (modules/govuk).

Licensify is a different beast to our other apps, and its puppet
manifest lives under modules/licensify. I've added a small bit of
logic to the App class to look for a 'puppet_url' property before
attempting to determine puppet URL via 'puppet_name'. This approach
is in line with the cascade approach we use for several properties
(e.g. puppet_name itself can be provided, or derived from the repo
name).