Create 'build' rake task
In the next commit, we'll define a deploy task which needs
the project to be built before it deploys. However, the current
'default' task runs a full test suite as part of the build -
something we won't need to do at the deploy stage, since it
will already have been tested.

This commit splits out the 'build' task from the 'default'
task, so that we can call it in isolation.