Create 'deploy' GitHub Action
Currently, a Jenkins job builds and deploys the developer docs to
an S3 bucket [using the 'deploy-to-s3' script][deploy].
It runs when a PR is merged, or between 9am and 7pm on weekdays
on a [timed cron job][job].

Historically we've used the S3 bucket in order to be proxied
through us for the TLS certificate, but we can now point the
CNAME to GitHub pages for hosting, which is one less moving part
to think about (two if you count both Jenkins _and_ S3), and
will cut some costs.

Step 1 is to get gh-pages building as regularly as the current
Jenkins builds. The next step will be to point
https://docs.publishing.service.gov.uk/ to GitHub Pages. The final
step will be to remove the legacy code and infrastructure.

Trello: https://trello.com/c/nEP3URQF/171-move-away-from-s3-bucket-to-github-pages-for-hosting-this-will-cut-some-costs-and-is-one-less-moving-part-to-think-about-from-ke

[deploy]: https://github.com/alphagov/govuk-puppet/blob/7f8de3c9ea11697b054aa644d1dec8d8071bb8b1/modules/govuk_jenkins/templates/jobs/deploy_docs.yaml.erb#L29
[job]: https://github.com/alphagov/govuk-puppet/blob/7f8de3c9ea11697b054aa644d1dec8d8071bb8b1/modules/govuk_jenkins/templates/jobs/deploy_docs.yaml.erb#L22