Create script to update gh-pages branch
This script can either be run locally or triggered in a
GitHub Action or inside the hourly Jenkins cron task.

Next we'll need to [update the Jenkins job][job] to call
`update-github-pages` in addition to `deploy-to-s3`. When
we're happy that `gh-pages` is deploying as regularly as
the S3 version, and happy that it works, we can work on
pointing https://docs.publishing.service.gov.uk/ to GitHub
Pages, and then remove the S3 bucket and associated code.

In the long term, we'll probably want to [use GitHub Actions
for deployment][action], which would let us kill off the
Jenkins job altogether. This is dependant on having use
of secrets signed off.

Trello: https://trello.com/c/nEP3URQF/171-host-the-dev-docs-on-github-pages

[action]: https://github.com/alphagov/govuk-developer-docs/pull/2656
[job]: https://github.com/alphagov/govuk-puppet/blob/7f8de3c9ea11697b054aa644d1dec8d8071bb8b1/modules/govuk_jenkins/templates/jobs/deploy_docs.yaml.erb#L29