Add new doc to describe auto security updates
Previously this was mixed up with rebooting guidance. Although
the original doc had guidance about choosing when to reboot, this
came from a legacy import [1] and it's unclear if we really need
to check the details of updates that have already been applied.

Note: the original "/var/log/unattended-upgrades" path suggestion
was wrong; the new doc has a corrected version.

[1]: https://github.com/alphagov/govuk-developer-docs/commit/5547bb99c17bbf9acc5841272596d585a61018e6