Move Google spreadsheet reference earlier in the documentation
We now have a number of keys documented in the spreadsheet, and the IAM
roles have been around long enough that we're starting to re-rotate some keys
we've already rotated in the past. It therefore makes sense to point the dev
to the spreadsheet in the first instance, to use it as a guide.