Run Vale markdown linting with GitHub Actions
Current implementation is a pre-commit hook which relies on the
developer to install Vale on their local machine. I'd be surprised
if every dev has done this, as I hadn't done so myself until
recently. Moreover, devs often update docs via the GitHub UI, which
would bypass any markdown linting checks.

Vale used to be set up to run on Jenkins, but has been lost
since we moved to GitHub Actions, so this attempts to correct
that. See:

https://github.com/alphagov/govuk-developer-docs/commit/fd67db213786389798fc73d58f695f6c17f0ac42#diff-58231b16fdee45a03a4ee3cf94a9f2c3R42