Install mdl (markdownlint/markdownlint)
We have quite inconsistent markdown throughout the Developer
Docs, and some mistakes can be caught with linting. For example,
some URLs written without any markdown links (`[]()`) don't
render as clickable links, and need to be encased in `<>`.

middleman 4.3.7 depends on Kramdown v1.2, but as a code
comment says, it's an 'optional middleman dependency,
included [only] for tests'.
https://github.com/middleman/middleman/blob/186bfb9c266b0ffd3f0824a33c614fe9afd3cf5e/Gemfile#L22-L25

mdl depends on Kramdown v2, so I had to manually bump the
dependency version in Gemfile.lock. I re-bundled and ran the
tests and it doesn't appear to have broken anything."