Add markdown linting rules
I've looked at the output of:
`bundle exec mdl --ignore-front-matter --show-aliases source`

And a couple of the most common errors can be disabled for
the reasons outlined in the comments.

Everything else thus far seems like a legitimate error
which should be fixed. I'll tackle these in the commits
that follow.