Fix markdown linting error: 'no-inline-html'
Unfortunately, this rule is slightly buggy, in that it treats
some URLs as HTML. We also need to embed an iframe in one of our
docs, so can't disable the rule altogether.

Until inline ignore is implemented (https://github.com/markdownlint/markdownlint/issues/16),
we're stuck disabling the rule as a whole. I've fixed what I can
manually.