Clarify that private repo READMEs aren't ingested
This would be a potential security risk if it did.
By design, calling 'application.readme' on a private repo
returns 'nil', which prior to this commit would have
rendered an empty div. Now we have an explicit message
explaining what is going on and why, which helps to
document the (intentional) behaviour.