Improve documentation on fact check alert
This includes some more explicit steps on what 2nd line can do when they
see this alert so hopefully it will be easier for them to handle. We're
doing this as an experiement as we don't believe the alert will fire
enough for us to need to put in place some automated systems, but if it
does end up being a problem we will revisit this decision.