Change CNAME value for transitioning sites.
This was discovered as part of a piece of 2ndline work to allow a
department to register geography.gov.uk and then redirect it to their
organistaion page.

The old CNAME appears to have been used when transitions onto GOV.UK
were a frequent occurange. If used now the Department's URL will just
redirect to a fastly error page.

However at some point we've switched over to this bounder-cdn. This PR
updates that in the docs so we won't use the old value again in future!