Determine README URL by checking 'default_branch' of repo
Currently all repos have a main branch of 'master', but in the near
future we expect this to change. We won't necessarily be in a place
where ALL repos have had their default branch renamed, so we need
to build in some conditional logic.

Introducing the logic check means making an indirect call to
'all_alphagov_repos' via 'repo', and this 1) checks the local cache,
then 2) makes a network request. We're already doing some network
stubbing and cache resetting in the '#repo' tests further up in the
file, and we need to take this a step further in these new tests.

I'm not very happy that I've had to use the same 'some-repo'
name in the '#readme' tests. However, every time I attempt to
use randomised repo names (or even a consistent, but different name
to 'some-repo'), I get errors around that repo not being found in
the cache.

'CACHE.clear' doesn't seem to help - but it _does_ seem to work if
I only ever work with the same cache key name (i.e. 'some-repo').
I've attempted a few things like 'config.before :each do {}' and
calling 'Rails.cache.clear' but that causes an error (as this isn't
a Rails app). The kind of cache we're using appears to be
'ActiveSupport::Cache::FileStore' (based on a 'pry') but I haven't
figured out how I can easily clear it between test runs.

NB I've also taken to running 'rm -rf .cache' prior to running
tests. It would be great if tests weren't subject to caching.