Update zendesk.html.md
In 197ff4b655be363c1833bfaa878f8397aff28776 I changed this from 5 days to 4 days as
this was the macro that I saw in Zendesk. However, that turns out to be a separate macro
and GOV.UK is still expected to follow the 5 day rule.

Also the markdown escaping from that commit did not work, so I've wrapped inside
backticks to preserve the `--`.