Update rabbitmq doc
This updates the main RabbitMQ documentation. On improving this doc I've
moved the `Nomenculture` section from the README in
[govuk_message_queue_consumer] into source/manual/rabbitmq.html.md as i
felt it was more at home here. I've linked to it from the consumer gem
so the information is still accessible there.

The original discussion around the decisions made in this doc can be
found [here][original_pr].

Related PR:
[govuk_message_queue_consumer]

Trello:
https://trello.com/c/0OHZT3cq/458-investigate-rabbitmq-and-improve-the-documentation

[govuk_message_queue_consumer]: https://github.com/alphagov/govuk_message_queue_consumer/pull/67
[original_pr]: https://github.com/alphagov/govuk-developer-docs/pull/2733