Remove app dependencies
This was added in 80b672a5d6a81334b95a1fbbdaf2d55d24cf6afe with very little context
('bla'), and appears to be required by other files in that commit (AppDependencyGenerator,
AppDiagramGenerator, etc), which were once used to generate diagrams.

Those generators and diagrams no longer exist in the dev docs, but for some reason
the application dependencies config still does. This isn't ordinarily a bad thing,
but if it isn't being used, it's not serving any useful purpose as documentation,
and risks falling out of date. It also looks somewhat buggy in the first place
(notice that service-manual-publisher has a dependency of... service-manual-publisher).

Let's just delete this unused config to make the config file simpler.