Update longstanding Alerts documentation
Spent some time testing each of these alert docs such that I know the
content within them are valid. Testing methodology involved running
any safe associated commands to look for errors that don't actually
exist (right now) because we're not seeing the alerts, checking links
and processes, etc.

We can't wait for these alerts to actually trigger to verify this
documentation so I did the closest next-best-thing I could think of and
tested them as if there was a valid alert I was trying to solve - within
reason.