Add 'last updated' banner to external docs (based on commit SHA)
Unlike the 'internal' docs that live in govuk-developer-docs,
these externally pulled in docs have no 'frontmatter' (YAML)
describing their state. All we have to go on is the commit history
itself.

This commit creates a 'latest_commit' property that is proxied
through to the frontend via ProxyPages. We use this to show
the 'last updated' value and link to the SHA of the commit so
that the viewer can see the latest changes.