Remove outdated comment
It seems at one point we had to write our own string to ID, as
the default anchoring of headings in Redcarpet used a different
algorithm to GitHub, meaning internal links were broken in places.
The most recent commit where that was relevant is here:
https://github.com/alphagov/govuk-developer-docs/commit/77f176ff84f4059af86d75ae3b09ade622f46cdf#diff-e1ce67e87f717ed2945410227009580fL8-L18

Since then, our need for a predictable string-to-id algorithm has
spread to multiple places. We have a 'breadcrumb' at the top of
each page, which links to the relevant section of the manual.html
page, which requires that we convert the section title to an ID
reliably.

We've now switched from Redcarpet to Kramdown, so the original
reason for rolling our own algorithm is negated, but we still need
to keep the string-to-id code as it's now referenced in several
places. It's not doing anything hugely complex and there's little
value in trying to remove it.

There is, however, value in removing the comment, which no longer
carries the significance it once did.