Rewrite Smokey alert doc to be more precise
Previously this doc contained several vague statements indicating
a lack of understanding about the cause of a problem. We should
try to spend our effort on solutions we have confidence will work,
rather than just trying random things. This rewrites the doc to
provide a clearer path for in-depth debugging.