Add 'last updated' banner to internal docs (based on commit SHA)
Unlike the external docs which are pulled in via [middleman proxy],
the internal docs are rendered 'natively' by middleman and don't
expose any hooks for adding to its frontmatter. Short of hard
coding each SHA and timestamp at the top of each doc's YAML file,
I've had to write a helper.

The helper examines the `current_page` and runs local git commands
against the source file to establish the SHA and timestamp of the
latest commit for the file.

[middleman proxy]: https://middlemanapp.com/advanced/dynamic-pages/