Defend against repos which are missing a README.md
GitHubRepoFetcher returns 'nil' if a README.md is missing from
the requested repo. But when we pass it to ExternalDoc.parse,
it doesn't know what to do with 'nil' and the build fails.

We're not seeing errors on CI as the README inclusion isn't
working at the moment (no ENV['GITHUB_TOKEN'] detected),
but it is failing locally on
https://github.com/alphagov/ckanext-datagovuk
(this has a README.rst, rather than a README.md).