Add timezone configuration to Rails conventions
Using London as the Rails timezone is a convention that has been used across
many GOV.UK apps going as far back as 2011 [1]. However this convention
does not seem to either be documented or used universally [2], so I felt
it was worthy of an inclusion.

This approach of using a local timezone compared to UTC does seem
inconsistent to the timing of our servers (UTC) and common Rails
recommendations [3]. However, this convention is long embedded into
GOV.UK Rails apps and it seems to be rare to imagine scenarios where
we'd not want London time to be the default for user interfaces.

[1]: https://github.com/alphagov/whitehall/commit/9cf416eaac95dc7dbb169b162d0a17b5c1976ac2
[2]: https://github.com/alphagov/collections/pull/1985/commits/35e3cb4850ea72b43f64dd5b207d4f3c4369ad74
[3]: https://thoughtbot.com/blog/its-about-time-zones