Document the Gemfile organisation convention
Over GOV.UK's lifespan Ruby Gemfiles have been organised and defined in
various ways with some gems defined as a particular fixed version
"1.0.1", some a range of minor releases "~> 1.5.2", others a range of
major versions "~> 2.0". This adds unnecessary thought and time every
time someone has to edit a Gemfile.

Luckily, the consulting team we brought in to help us upgrade all apps
to Rails 6 have established a convention across Gemfiles of all the apps
they are updating [1] [2]. Which means for the first time GOV.UK now has an
established convention for Gemfiles.

This documentation aims to capture and document this convention, with an
intention that this change can be publicised to make others in the
GOV.UK development community aware of it.

[1]: https://github.com/alphagov/travel-advice-publisher/blob/4b46266f76a17dbbbd29a0ed38ffd909eb3eb637/Gemfile
[2]: https://github.com/alphagov/service-manual-frontend/pull/749#pullrequestreview-447953266