Document licence and readme for a new Rails app
We used to have templates for these files that were applied
automatically by https://github.com/alphagov/govuk-rails-app-template.
This project was replaced by this document however we lost the
documentation of the conventions for the software licence and README.md.

I've put the licence in a details element so it's contents can be
progressively disclosed. For this to work with Kramdown we need to tell
Kramdown to use block HTML parsing, hence the markdown attribute. See
https://kramdown.gettalong.org/syntax.html#html-blocks for further
details.