Add documentation on configuring linting
I wanted to bring the linting documentation on "Set up a new Rails
application" to reflect our current linting practices. However, I found
that the steps to do so were to verbose to fit in that documentation so
I broke it out into a separate one.

It is a little forward thinking by specifying stylelint-config-gds as
we're early days towards adopting that. However I think the choice of
that is still relevant as that is what any new projects should use.