Fetch full repo history when cloning via GitHub Actions
We ran into an [issue][] whereby the 'last updated' logic worked
locally but not on production. All 'local' pages were showing
today's date rather than the date they were last updated.

I hypothesised that `current_page` evaluated to an empty string,
giving the `source_file` value of `source/` for all local docs,
and thereby grabbing the latest commit for any file in that folder.
However, I now believe the issue lies with the way the repo was
cloned using GitHub Actions.

By [default][], only a single commit is fetched when cloning a
repo using the 'actions/checkout' action. This is for efficiency
reasons. But it means that we can't reliably query the commit
history for particular files, within the action itself - everything
is wrongly attributed to the single latest commit.

By adding 'fetch-depth: 0', we set an infinite fetch depth. In
other words, the entire repository history is cloned, so we should
get accurate results when querying the commit SHA and timestamp of
particular files.

[default]: https://github.com/actions/checkout
[issue]: https://github.com/alphagov/govuk-developer-docs/pull/2784#issuecomment-717786517