Update "Review a page in this manual" guidance
This removes references to the old 'Daniel the Manual Spaniel'
workflow, and documents how someone might want to update the
'Last updated' value.

We've deliberately not documented how to 'bump' a 'last updated'
date for documents that require no edits, as we don't want to
pre-empt something that may not be an issue, and don't want to
encourage an approach that could one day lead to inconsistencies.

For full disclosure, we were hoping to be able to use empty
commits for that task, but [this isn't possible on a per-file
basis](https://stackoverflow.com/questions/52315208/how-to-add-an-empty-git-commit-message-to-a-specific-file/52318584#52318584).
We could encourage people to add a `last_reviewed_on: YYYY-MM-DD`
property as before, but this would only be applicable to docs
hosted in govuk-developer-docs; we would still need to find a
workaround for externally hosted docs. Alternatively, we could
advocate for a HTML comment approach, which won't get rendered
to the reader.