Display 'last_reviewed_on' field if it exists
We had intended to encourage no-op commits on a per-file basis
to show the file has been reviewed, but this [doesn't seem
possible](https://stackoverflow.com/questions/52315208/how-to-add-an-empty-git-commit-message-to-a-specific-file/52318584#52318584).
So, we've decided to keep the `last_reviewed_on` field, which
people will be encouraged to update as and when they review
documentation.

In the event that no `last_reviewed_on` field exists, this will
fall back to checking the timestamp of the latest commit made to
the file in question.