Update "Review a page in this manual" guidance
This removes references to the old 'Daniel the Manual Spaniel'
workflow.

What isn't documented is how one might update the 'last updated'
value for an external document (which is based on its most recent
commit). We don't anticipate this being much of an issue, as
external documentation isn't subject to the same turnover of
reviewing and updating. If people are keen on updating the last
updated value of an external doc, there is usually something there
that can be edited. Alternatively they could create a new commit
by adding a HTML comment to a document, which doesn't get rendered
to the reader.