Display warning if document more than 6 months old
The 6 month cut-off is an arbitrary choice, but many of the
'review_in' values in #2837 were for 6 months, so it seems a good
indicator of age.

This commit uses the warning-text [component][] to highlight old
documents, which gives a clear visual clue to the reader when a
document might be out of date. It should hopefully encourage
readers to take ownership and review/update the document.

[component]: https://design-system.service.gov.uk/components/warning-text/

The styles are automatically pulled in from the tech docs gem:
https://github.com/alphagov/tech-docs-gem/blob/master/lib/assets/stylesheets/_govuk_tech_docs.scss#L17

If we wanted full govuk_publishing_components support, we could
pull in the gem and then render components like so:

```erb
<%= GovukPublishingComponents.render(
  "govuk_publishing_components/components/warning_text",
  text: "This document has not been updated for a while now. It may be out of date."
) %>
```

However, after some [discussion][] we deemed this not worth the
overhead of pulling in a gem that we then have to keep up to date.

[discussion]: https://github.com/alphagov/govuk-developer-docs/pull/2842#issuecomment-723129931